+++ src/main/groovy/ru/vyarus/gradle/plugin/python/cmd/docker/DockerConfig.groovy
@@ -19,5 +19,7 @@ class DockerConfig {
     boolean windows
     String image
     boolean exclusive
+    // would always be false for mac and win
+    boolean useHostNetwork
     Set<String> ports
 }